Move invalid msg test [skip ci]
