--- types/spectest/tests/partialsigmessage/invalid_messages.go
@@ -1,6 +1,0 @@
-package partialsigmessage
-
-// InvalidMessages tests PartialSignatureMessages with > 0 invalid message
-func InvalidMessages() *MsgSpecTest {
-	panic("implement")
-}
